HighDimMixedModels.L_diag_update! — Method

Update of coordinate s of L for diagonal covariance structure

ARGUMENTS

L :: A vector of parameters which will be updated by the function

s :: The coordinate of L that is being updated (number between 1 and length(L))

HighDimMixedModels.L_ident_update — Method

Update of L for identity covariance structure

HighDimMixedModels.L_sym_update! — Method

Update of L for general symmetric positive definite covariance structure ARGUMENTS

L :: A lower triangular matrix of parameters which will be updated by the function

coords :: Tuple representing the coordinates of the entry of L that is being updated

HighDimMixedModels.armijo! — Method

Armijo Rule

HighDimMixedModels.cov_start — Method

Finds an initial value for the variance and covariance parameters

ARGUMENTS

XGgrp :: Vector of fixed effect design matrices for each group

ygrp :: Vector of vector of responses for each group

Zgrp :: Vector of random effects design matrix for each group

β :: Initial iterate for fixed effect parameter vector (computed with Lasso ignoring group structure)

OUTPUT

Assuming β is true fixed effect vector, MLE estimate of scalar L and scalar σ² as tuple (L, σ²)

HighDimMixedModels.get_cost — Function

Calculates the objective function

HighDimMixedModels.get_negll — Method

Calculates the negative log-likelihod -l(ϕ̃) = -l(β, θ, σ²) = .5(Ntot*log(2π) + log|V| + (y-xβ)'V⁻¹(y-Xβ))

ARGUMENTS

invVgrp :: Vector of length the number of groups, each of whose elements is the precision matrix of the responses within a group

ygrp :: Vector of vector of responses for each group

X :: Vector of fixed effect design matrices for each group

β :: Fixed effects

OUTPUT

Value of the negative log-likelihood

HighDimMixedModels.get_scad — Function

Calculates the SCAD penalty

HighDimMixedModels.hessian_diag! — Method

Calculates active_set entries of the diagonal of Hessian matrix for fixed effect parameters

HighDimMixedModels.invV! — Method

Updates precision matrices of the responses, by group

ARGUMENTS

invVgrp :: Container for precision matrices of the responses, by group

Zgrp :: Container of random effects design matrices, by group

L :: Parameters for random effect covariance matrix (can be scalar, vector, or lower triangular matrix)

σ² :: Variance of error

OUTPUT

invVgrp :: List of length the number of groups, each of whose elements is the covariance matrix of the responses within a group

HighDimMixedModels.scad_dir — Method

Calculates descent direction with SCAD penalty

HighDimMixedModels.scad_solution — Method

Gets analytical solution for CGD iterate with SCAD penalty when the Hessian hasn't been truncated

HighDimMixedModels.soft_thresh — Method

Soft Threshold

HighDimMixedModels.special_quad — Method

Calculates (y-ỹ)'(invV)X[:,j], where ỹ are the fitted values if we ignored the jth column i.e. XG[:,Not(j)]β[Not(j)] To improve perforamce, we calculate ỹ with the entire dataset We then split into groups and calculate (y-ỹ)'(invV)*X[:,j] for each group

HighDimMixedModels.σ²update — Method

Update of σ²