HighDimMixedModels.Control — Type

Control Provides hyperparameters for the coordinate descent algorithm.

Fields

tol: Small positive number, default is 1e-4, providing convergence tolerance

seed :: Random seed, default 770. Note that the only randomness in the algorithm is during the initialization of fixed effect parameters (for the data splits in the cross validation)

trace :: Integer, default 2. 1 prints no output, 2 prints issues, and 3 prints issues and the objective function values during the algorithm

max_iter :: Integer, default 1000, giving maximum number of iterations in the coordinate gradient descent.

max_armijo :: Integer, default 20, giving the maximum number of steps in the Armijo algorithm. If the maximum is reached, algorithm doesn't update current coordinate and proceeds to the next coordinate

actnum :: Integer, default 5. We will only update all of the fixed effect parameters every `actnum` iterations. Otherwise, we update only the parameters in the current active set.

a₀ :: a₀ in the Armijo step, default 1.0. See Schelldorfer et al. (2010) for details about this and the next five fields.

δ :: δ in the Armijo step, default 0.1.

ρ :: ρ in the Armijo step, default 0.001.

γ :: γ in the Armijo step, default 0.0.

lower :: Lower bound for the Hessian, default 1e-6.

upper :: Upper bound for the Hessian, default 1e8.

var_int :: Tuple with bounds of interval on which to optimize when updating a diagonal entry of L, default (0, 100). See Optim.jl in section "minimizing a univariate function on a bounded interval"

cov_int :: Tuple with bounds of interval on which to optimize the when updating a non-diagonal entry of L, default (-50, 50). See Optim.jl in section "minimizing a univariate function on a bounded interval"

optimize_method :: Symbol denoting method for performing the univariate optimization, either :Brent or :GoldenSection, default is :Brent

thres :: If an update of an entry of L or of σ² would be smaller in absolute value than thres, the parameter is set to 0

HighDimMixedModels.HDMModel — Type

Stores the results of a fitted model

HighDimMixedModels.hdmm — Function

hdmm(X::Matrix{<:Real}, G::Matrix{<:Real}, y::Vector{<:Real}, 
    grp::Vector{<:Union{String, Int64}}, Z::Matrix{<:Real}=X;
    standardize=true, penalty::String="scad", λ::Real=10.0, scada::Real=3.7, 
    wts::Union{Vector, Nothing}=nothing, init_coef::Union{Vector,Nothing}=nothing, 
    ψstr::String="diag", control::Control=Control())

Main function for the package. Fits a penalized linear mixed effect model using the coordinate gradient descent algorithm.

Arguments

Positional:

X: Low dimensional (N by q) design matrix for unpenalized fixed effects (first column must be 1's to fit intercept)

G: High dimensional (N by p) design matrix for penalized fixed effects (should not include column of 1's)

y: Response vector

grp: Vector of either strings or integers with group assignments of each observation

Z: Design matrix (N by m) for random effects, defaults to equal X

Keyword:

penalty: String, one of "scad" (default) or "lasso"

λ: Positive number, default=10, providing the regularization parameter for the penalty

scada: Positive number, default=3.7, providing the extra tuning parameter for the SCAD penalty (ignored if penalty is "lasso")

standardize: Boolean, deault=true, whether to standardize the columns of all design matrices before performing coordinate descent. Note that the value of λ should be chosen accordingly. Results will always be returned on the original scale

wts: Vector of length p, default is vector of 1's. The penalty on covariate j will be λ/wⱼ, so this argument is useful if you want to penalize some covariates more than others.

ψstr: String, one of "diag" (default), "ident", or "sym", specifying covariance structure of the random effects

init_coef: Named tuple of form (β, L, σ²), where

β is a vector of length p + q providing an initial estimate of the fixed effect coefficients

L is the Cholesky factor of the random effect covariance matrix, and is represented as

a scalar if ψstr="ident"

a vector of length m if ψstr="diag"

a lower triangular matrix of size m by m if ψstr="sym"

σ² is a scalar providing an initial estimate of the noise variance

See below for how these inital parameter estimates are calculated if not provided.

control: Custom struct with fields for hyperparameters of the algorithm, defaults are in documentation of Control struct

If the init_coef argument is not specified, we perform the following steps to obtain the initial parameters to feed to the coordinate descent algorithm:

First, a LASSO (with λ chosen using cross validation) that ignores random effects is performed to estimate the fixed effect parameters.

Then, L, assumed for the moment to be a scalar, and σ² are estimated to maximize the likelihood given the estimated fixed effect parameters.

Finally, if ψstr is "diag" or "sym", the scalar L is converted to a vector or matrix by repeating the scalar or filling the diagonal of a matrix with the scalar, respectively.

Returns

Fitted model object of type HDMModel with fields for all relevant information about the model fit.

StatsAPI.aic — Method

aic(fit::HDMModel)

The Akaike Information Criterion is equal to the deviance plus 2 times the number of parameters in the model

StatsAPI.bic — Method

bic(fit::HDMModel)

The Bayesian Information Criterion is equal to the deviance plus the log of the number of observations times the number of parameters in the model

StatsAPI.coef — Method

coef(fit::HDMMModel)

Retrieves all estimated fixed effect coefficients

StatsAPI.coeftable — Function

coeftable(fit::HDMModel, names::Vector{String}=string.(1:length(fit.fixef)))

Returns a table of the selected coefficients, i.e. those not set to 0, from the model.

Arguments

fit::HDMModel: A fitted model.

names::Vector{String}: Names of the all the coefficients in the model (not just those selected), defaults to integer names

Returns

A StatsBase.CoefTable object.

StatsAPI.deviance — Method

deviance(fit::HDMModel)

-2*loglikelihood of the model at the estimated parameters

StatsAPI.fitted — Method

fitted(fit::HDMModel)

Accounts for the random effects in generating predictions

StatsAPI.loglikelihood — Method

loglikelihood(fit::HDMModel)

Log-likelihood of the model at the estimated parameters

StatsAPI.nobs — Method

nobs(fit::HDMModel)

Number of observations used in fitting the model

StatsAPI.residuals — Method

residuals(fit::HDMModel)

Accounts for the random effects in generating predictions